fix: casting float to int before saving it as a BIGINT
There was a 500 error when blown kisses hit the target:

```
web-1  | NOTICE: PHP message: SQLSTATE[22P02]: Invalid text representation: 7 ERROR:  invalid input syntax for type bigint: "68103.904791222"
web-1  | CONTEXT:  unnamed portal parameter $1 = '...'
web-1  | 127.0.0.1 -  26/Nov/2025:14:06:37 +0000 "GET /index.php" 500
```

+++ apps/monolith/phpincludes/fctIndex.php
@@ -202,7 +202,7 @@ function AjouterScore($idScore, $valeur): void
 
     $donnees_info = $stmt->fetch();
     $stmt = $pdo->prepare('UPDATE membres SET score = :score WHERE id = :id');
-    $stmt->execute(['score' => $donnees_info['score'] + $valeur, 'id' => $idScore]);
+    $stmt->execute(['score' => (int) ($donnees_info['score'] + $valeur), 'id' => $idScore]);
 }
 
 function formaterNombre($nombre): string
